tracker/pyproject.toml
[tool.poetry]
name = "tracker"
version = "0.1.0"
description = ""
authors = ["Quint van Dijk <quintvdijk@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
alive-progress = "^3.1.4"
websockets = "^11.0.3"
pillow = "^10.0.0"
playsound = "^1.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
